Moved more code around
